+++ R/write.mplus.R
@@ -1,18 +1,22 @@
 #' Write Mplus Data File
 #'
-#' This function writes a matrix or data frame to a tab-delimited file without variable names and
-#' a text file with variable names. Only numeric values are allowed, missing data will be coded
-#' as a single numeric value.
+#' This function writes a matrix or data frame to a tab-delimited file without variable names,
+#' a Mplus input template, and a text file with variable names. Note that only numeric variables
+#' are allowed, i.e., non-numeric variables will be removed from the data set. Missing data
+#' will be coded as a single numeric value.
 #'
-#' @param x           a matrix or data frame to be written to a tab-delimited file.
-#' @param file        a character string naming a file with or without the file extension '.dat',
-#'                    e.g., \code{"Mplus_Data.dat"} or \code{"Mplus_Data"}.
-#' @param var         logical: if \code{TRUE}, variable names are written in a text file named
-#'                    according to the argument\code{file} with the extension \code{_VARNAMES.txt}.
-#' @param print       logical: if \code{TRUE}, variable names are printed on the console.
-#' @param na          a numeric value or character string representing missing values (\code{NA})
-#'                    in the data set.
-#' @param check       logical: if \code{TRUE}, argument specification is checked.
+#' @param x      a matrix or data frame to be written to a tab-delimited file.
+#' @param file   a character string naming a file with or without the file extension '.dat',
+#'               e.g., \code{"Mplus_Data.dat"} or \code{"Mplus_Data"}.
+#' @param input  logical: if \code{TRUE} (default), Mplus input template is written in a text file named
+#'               according to the argument\code{file} with the extension \code{_INPUT.inp}.
+#' @param n.var  a numeric value indicating the number of variables in each line under \code{NAMES ARE}
+#'               in the the Mplus input template.
+#' @param var    logical: if \code{TRUE}, variable names are written in a text file named
+#'               according to the argument\code{file} with the extension \code{_VARNAMES.txt}.
+#' @param na     a numeric value or character string representing missing values (\code{NA})
+#'               in the data set.
+#' @param check  logical: if \code{TRUE}, argument specification is checked.
 #'
 #' @author
 #' Takuya Yanagida \email{takuya.yanagida@@univie.ac.at}
@@ -30,19 +34,16 @@
 #'
 #' @examples
 #' \dontrun{
-#' dat <- data.frame(id = 1:5,
-#'                   x = c(NA, 2, 1, 5, 6),
-#'                   y = c(5, 3, 6, 8, 2),
-#'                   z = c(2, 1, 1, NA, 4), stringsAsFactors = FALSE)
+#' # Write Mplus Data File and a Mplus input template
+#' write.mplus(mtcars)
 #'
-#' # Write Mplus Data File and a text file with variable names
-#' write.mplus(dat)
-#'
-#' # Write Mplus Data File "Data.dat" and a text file with variable name,
-#' # print variable names on the console, missing values coded with -999
-#' write.mplus(dat, file = "Data.dat", print = TRUE, na = -999)
+#' # Write Mplus Data File "mtcars.dat" and a Mplus input template "mtcars_INPUT.inp",
+#' # missing values coded with -999, 4 variables in each line under "NAMES ARE"
+#' # write variable names in a text file called "mtcars_VARNAMES.inp"
+#' write.mplus(mtcars, file = "mtcars.dat", n.var = 4, var = TRUE, na = -999)
 #' }
-write.mplus <- function(x, file = "Mplus_Data.dat", var = TRUE, print = FALSE, na = -99, check = TRUE) {
+write.mplus <- function(x, file = "Mplus_Data.dat", input = TRUE, n.var = 8, var = FALSE,
+                        na = -99, check = TRUE) {
 
   ####################################################################################
   # Data
@@ -68,6 +69,26 @@ write.mplus <- function(x, file = "Mplus_Data.dat", var = TRUE, print = FALSE, n
 
   x <- as.data.frame(x, stringsAsFactors = FALSE)
 
+  #......
+  # Numeric variables
+  x.numeric <- vapply(x, is.numeric, FUN.VALUE = logical(1))
+
+  if (isTRUE(any(!x.numeric))) {
+
+    x <- x[, x.numeric]
+
+    warning(paste0("Non-numeric variables were excluded from the data set: ",
+                   paste(names(which(!x.numeric)), collapse = ", ")), call. = FALSE)
+
+    if (isTRUE(ncol(x) == 0L)) {
+
+      stop("No variables left for the data set after excluding non-numeric variables.",
+           call. = FALSE)
+
+    }
+
+  }
+
   ####################################################################################
   # Input Check
 
@@ -84,21 +105,26 @@ write.mplus <- function(x, file = "Mplus_Data.dat", var = TRUE, print = FALSE, n
   if (isTRUE(check)) {
 
     #......
-    # Numeric variables
-    x.numeric <-  vapply(x, is.numeric, FUN.VALUE = logical(1))
+    # Check input 'input'
+    if (isTRUE(!is.logical(input))) {
 
-    if (isTRUE(any(!x.numeric))) {
+      stop("Please specify TRUE or FALSE for the argument 'input'.", call. = FALSE)
 
-      x <- x[, x.numeric]
+    }
 
-      warning(paste0("Non-numeric variables were excluded from the data set: ", paste(names(which(!x.numeric)), collapse = ", ")),
-              call. = FALSE)
+    #......
+    # Check input 'n.var'
+    if (isTRUE(n.var %% 1L != 0L || n.var < 0L)) {
 
-      if (isTRUE(ncol(x) == 0L)) {
+      stop("Specify a positive integer number for the argument 'n.var'.", call. = FALSE)
 
-        stop("No variables left for the data set after excluding non-numeric variables.", call. = FALSE)
+    }
 
-      }
+    #......
+    # Check input 'var'
+    if (isTRUE(!is.logical(var))) {
+
+      stop("Please specify TRUE or FALSE for the argument 'var'.", call. = FALSE)
 
     }
 
@@ -110,7 +136,8 @@ write.mplus <- function(x, file = "Mplus_Data.dat", var = TRUE, print = FALSE, n
 
       names(x) <- gsub("\\.", "_", names(x))
 
-      warning("Special character \".\" (dot) in the variable names were replaced with  \"_\" (underscore).", call. = FALSE)
+      warning("Special character \".\" (dot) in the variable names were replaced with  \"_\" (underscore).",
+              call. = FALSE)
 
     }
 
@@ -161,19 +188,70 @@ write.mplus <- function(x, file = "Mplus_Data.dat", var = TRUE, print = FALSE, n
 
   if (isTRUE(var)) {
 
-    file <- sub(names(which(vapply(sapply(c(".dat", ".txt", ".csv"), grep, file), length, FUN.VALUE = integer(1)) != 0L)),
-                "_VARNAMES.txt", file)
+    file.var <- sub(names(which(vapply(sapply(c(".dat", ".txt", ".csv"), grep, file), length, FUN.VALUE = integer(1)) != 0L)),
+                    "_VARNAMES.txt", file)
 
-    writeLines(paste(names(x), collapse = " "), con = file)
+    writeLines(paste(names(x), collapse = " "), con = file.var)
 
   }
 
   #-----------------------------------------
-  # Print variable names on console
+  # Save Mplus input
 
-  if (isTRUE(print)) {
+  if (isTRUE(input)) {
 
-    print(names(x))
+    # Length of variable names divided by n.var
+    var.n <- length(names(x)) / n.var
+
+    temp <- NULL
+    #....
+    # Length of variable names greater than n.var
+    if (isTRUE(floor(var.n) > 0L)) {
+
+      var.x <- c(0L, sapply(1L:floor(var.n), function(y) y*n.var))
+
+      for (i in seq_along(var.x)[-length(var.x)] ) {
+
+          temp <- c(temp,
+                    paste0("            ", paste(names(x)[(var.x[i] + 1L):var.x[i + 1L]], collapse = " "), "\n"))
+
+      }
+
+      # Add remaining variables
+      if (isTRUE(var.n %% 1L != 0L)) {
+
+        temp <- c(temp,
+                  paste0("            ", paste(names(x)[(var.x[length(var.x)] + 1L):length(names(x))], collapse = " "), ";\n"))
+
+      } else {
+
+        temp[length(temp)] <- sub("\n", ";\n", temp[length(temp)])
+
+      }
+
+    #....
+    # Length of variable names smaller than n.var
+    } else {
+
+      temp <- c(temp,
+                paste0("            ", paste(names(x), collapse = " "), ";\n"))
+
+    }
+
+    #....
+    # Mplus input template
+    temp <- paste0("DATA:       FILE IS ", file, ";\n\n",
+                   "VARIABLE:   NAMES ARE \n", paste(temp, collapse = ""), "\n",
+                   "            USEVARIABLES ARE ", ";\n\n",
+                   "            MISSING IS ALL(", na, ");\n\n",
+                   "ANALYSIS:   \n\n",
+                   "MODEL:      \n\n",
+                   "OUTPUT:     \n")
+
+    file.inp <- sub(names(which(vapply(sapply(c(".dat", ".txt", ".csv"), grep, file), length, FUN.VALUE = integer(1)) != 0L)),
+                    "_INPUT.inp", file)
+
+    writeLines(temp, con = file.inp)
 
   }